requirements.txt
requests==2.31.0
flask>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
gunicorn>=20.1.0
httpx>=0.24.0